Move Pairer to enable reuse
This commit moves the helper Pairer out of reduce.rs and into
mod.rs, as this is a better location for reuse by other rendering
modules. The commit comprises code movement ontly.
